Fix repo-index
Sed syntax was not working in build.sh (not sure why exactly, since it was successful if I copied and pasted the sed commands and ran on the commandline with manual variable substitution).  Anyway, these changes fix the issue while simplifying the sed syntax and in turn, entering a date for the repo index no longer requires escaping of forwardslashes.